bug: 30550
Blob size is a long, now matching MailboxBlob.

http://bugzilla.zimbra.com/show_bug.cgi?id=30550

+++ ZimbraOffline/src/java/com/zimbra/cs/mailbox/InitialSync.java
@@ -1151,7 +1151,7 @@ public class InitialSync {
                 pm = new ParsedMessage(blob.getFile(), received * 1000L, false);
                 digest = blob.getDigest();
                 pm.setRawDigest(digest);
-                size = blob.getRawSize();
+                size = (int) blob.getRawSize();
             }
         } catch (IOException e) {
             throw MailServiceException.MESSAGE_PARSE_ERROR(e);
